+++ guizero/guizerocontroleventcore.py
@@ -1,5 +1,3 @@
-from .events import MouseEvents, KeyEvents
-
 
 class GuiZeroControlEventCore:
     # Bind to Events
@@ -63,3 +61,26 @@ class GuiZeroControlEventCore:
     def on_focus_out(self, func, add=""):
         """ Delegates the action when the control no longer has focus """
         self.bind(KeyEvents.FOCUS_OUT, func, add)
+
+class KeyEvents:
+    """ Keyboard events to bind a function to control action """
+    KEY_UP = "<KeyRelease>"
+    KEY_PRESS = "<KeyPress>"
+    FOCUS_IN = "<FocusIn>"
+    FOCUS_OUT = "<FocusOut>"
+
+
+class MouseEvents:
+    """ Mouse events to bind a function to control action """
+    MOUSE_PRESS_LEFT_BUTTON = "<Button-1>"
+    MOUSE_PRESS_MIDDLE_BUTTON = "<Button-2>"
+    MOUSE_PRESS_RIGHT_BUTTON = "<Button-3>"
+    MOUSE_MOVE_HOLD_LEFT_BUTTON = "<B1-Motion>"
+    MOUSE_MOVE_HOLD_MIDDLE_BUTTON = "<B2-Motion>"
+    MOUSE_MOVE_HOLD_RIGHT_BUTTON = "<B3-Motion>"
+    MOUSE_RELEASE_LEFT_BUTTON = "<ButtonRelease-1>"
+    MOUSE_RELEASE_MIDDLE_BUTTON = "<ButtonRelease-2>"
+    MOUSE_RELEASE_RIGHT_BUTTON = "<ButtonRelease-3>"
+    MOUSE_ENTER = "<Enter>"
+    MOUSE_LEAVE = "<Leave>"
+    MOUSE_WHEEL = "<MouseWheel>"
